Fix test config
