+++ tests/config.py
@@ -1,0 +1,3 @@
+PRESERVE_CONTEXT_ON_EXCEPTION = False
+SQLALCHEMY_DATABASE_URI = "sqlite://"
+TESTING = True
